rename directory modules to utilities and add cloudinary sdk
